+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Support old versions of react (pre v16.14)
+- Prefer idiomatic phrasing by default (use `numeric: 'always'` for the old behaviour)
 
 ## 2.0.0 (2024-02-09)